change main portion
https://wsoteam.atlassian.net/browse/TN-17054
https://wsoteam.atlassian.net/browse/TN-17053

+++ app/src/main/java/com/wsoteam/diet/common/networking/food/POJO/Result.java
@@ -80,7 +80,7 @@ public class Result implements ISearchResult, Serializable {
         return portion;
     }
 
-    public void setPortion(int portion) {
+    public void setPortion(double portion) {
         this.portion = portion;
     }
